Split `allocator` module into `phys` and `heap`
Containing two different kinds of codes is bad.
